Moving StepAttemptUpdate Mappers so the they live with the DTOs they extend (StepAttemotDetail and StepAttemptForm DTOs)
